Start with some pseudo code and outlining
- Start an empty function (naming things is hard) and put in some pseudo code
after looking at it for a bit I sense sort of an abstracted fizzbuzz pattern
coming up with a slight twist. I know simplest gotcha pieces of fizzbuzz is
starting with the most complex condition first, to catch all pieces of it.
- I think I'm going to start with the days and output. No other conditions.
Then same for hours, then minutes, then I'll focus on the zero stripping and
the string interpolation. Then refactor.
- Make it work, make it right, make it better.